+++ src/notion/loadPage.ts
@@ -5,11 +5,11 @@ import {
   NotionPageDescription,
   NotionPageImage,
   NotionPageLinkedPage,
-  NotionPageText,
+  NotionMeta,
 } from '../types/notion';
 
-import parseMetaBlock from './parser/parseMetaBlock';
 import notionPageTextToString from './parser/notionPageTextToString';
+import parseMetaText from './parser/parseMetaText';
 
 function getPropertyAsString(
   block: NotionPageBlock,
@@ -35,17 +35,6 @@ function getAttributeAsString(
   return att.value;
 }
 
-function getPropertyText(
-  block: NotionPageBlock,
-  propName: string,
-): NotionPageText[] | null {
-  const property = block.properties.find(p => p.propName === propName);
-  if (!property) {
-    return null;
-  }
-  return property.value;
-}
-
 // loads a gatsby page
 export default async function loadPage(
   pageId: string,
@@ -53,6 +42,7 @@ export default async function loadPage(
   indexPage: number,
   notionLoader: NotionLoader,
   reporter: Reporter,
+  debug: boolean,
 ): Promise<NotionPageDescription> {
   // we load the given page
   await notionLoader.loadPage(pageId);
@@ -70,17 +60,15 @@ export default async function loadPage(
 
   const imageDescriptions: NotionPageImage[] = [];
   const linkedPages: NotionPageLinkedPage[] = [];
-  let hasMeta = false;
-  const meta: Record<string, string> = {};
-  let hasExcerpt = false;
-  let excerpt = '';
+  const meta: NotionMeta = {};
+  const metaParser = parseMetaText(meta);
 
-  // parse all the blocks retrived from notion
+  // parse all the blocks retrieved from notion
   for (const blockId of page.blockIds) {
     const block = notionLoader.getBlockById(blockId);
     if (!block) {
-      reporter.error(`could not retreieve para with id: ${blockId}`);
-      throw Error('error retrieving paragraph');
+      reporter.error(`could not retrieve block with id: ${blockId}`);
+      throw Error('error retrieving block in page');
     }
     switch (block.type) {
       case 'page':
@@ -90,26 +78,13 @@ export default async function loadPage(
         });
         break;
       case 'text':
-        // the first non empty text becomes the excerpt
-        if (!hasExcerpt) {
+        {
+          // for the text blocks, we parse them to see if they contain
+          // meta attributes, if not, we addf them as regular blocks
           const text = getPropertyAsString(block, 'title', '').trim();
-          if (text.length > 0) {
-            hasExcerpt = true;
-            excerpt = text;
-          }
-        }
-        break;
-      case 'quote':
-        if (!hasMeta) {
-          hasMeta = true;
-          const text = getPropertyText(block, 'title');
-          // try to parse the block as a meta information definition block
-          if (text) {
-            if (parseMetaBlock(text, meta)) {
-              // if we were able to parse the block, we change its type
-              // (so that it is not rendered as a quote)
-              block.type = '_meta';
-            }
+          if (metaParser(text)) {
+            // we change the type to meta to avoid the rendering of this text block
+            block.type = 'meta';
           }
         }
         break;
@@ -125,42 +100,8 @@ export default async function loadPage(
         // guess what... we ignore that one
         break;
       default:
-        // we keep the record by defaut
+        // we keep the block by defaut
         break;
-    }
-  }
-
-  // default meta value for the page
-  // can be overriden from the page, using a quote block
-  let slug = `${indexPage}`;
-  let createdAt = new Date().toISOString();
-  let isDraft = false;
-  const tags: string[] = [];
-
-  if (hasMeta) {
-    if (meta['slug']) {
-      slug = meta['slug'];
-    }
-    // TODO: parse date so that it becomes an actual
-    // date in GraphQL
-    if (meta['date']) {
-      createdAt = new Date(meta['date'] + ' Z').toJSON();
-    }
-    if (meta['draft']) {
-      const value = meta['draft'].toLowerCase();
-      if (value === 'false' || value === '0') {
-        isDraft = false;
-      } else {
-        isDraft = true;
-      }
-    }
-    // tags can contain a comma separated list of value
-    if (meta['tags']) {
-      const value = meta['tags'].trim();
-      value
-        .split(',')
-        .map(t => t.trim())
-        .forEach(t => tags.push(t));
     }
   }
 
@@ -168,16 +109,18 @@ export default async function loadPage(
     pageId,
     title: getPropertyAsString(page, 'title', ''),
     indexPage,
-    slug,
-    createdAt,
-    tags,
-    isDraft,
-    excerpt,
+    slug: meta.slug || `${indexPage}`,
+    createdAt: meta.date || new Date().toISOString(),
+    tags: meta.tags || [],
+    isDraft: !!meta.isDraft,
+    excerpt: meta.excerpt || '',
     pageIcon: getAttributeAsString(page, 'pageIcon', ''),
     blocks: [],
     images: imageDescriptions,
     linkedPages,
   };
+  // we return all the blocks
+  // TODO: as we already got those blocks above, we may want to build the list as we go
   notionLoader.getBlocks(item.blocks, rootPageId);
   return item;
 }